pyproject.toml
[project]
name = "aidebate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "google-generativeai>=0.8.4",
    "langchain-community>=0.3.19",
    "langchain>=0.3.20",
    "openai>=1.65.5",
    "fastapi>=0.110.0",
    "jinja2>=3.1.3",
    "uvicorn>=0.27.1",
    "python-multipart>=0.0.9",
    "psutil>=5.9.8",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.10",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
